pyproject.toml
[project]
name = "ragent"
version = "0.1.0"
description = "A RAG pipeline using Azure OpenAI and LangChain"
authors = [
    { name = "Martin", email = "martin@example.com" }
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.10",
    "langgraph>=0.0.15",
    "chromadb>=0.4.22",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.8.0",
]

[project.scripts]
ragent = "ragent.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ragent"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto" 